# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel Separated User Auth(JWT) Example",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/auth/company/login": {
            "post": {
                "tags": [
                    "Company Authentication"
                ],
                "summary": "User login",
                "description": "Authenticate company user and return JWT tokens",
                "operationId": "f0d262b157a0075a92e084aca7f9115d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "company@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token pair response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "refresh": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/company/refresh": {
            "post": {
                "tags": [
                    "Company Authentication"
                ],
                "summary": "Refresh JWT token",
                "description": "Refresh company user authentication token",
                "operationId": "392cadbbe8972303872dd761b76d7be5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "token": {}
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "access": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "refresh": {
                                            "properties": {
                                                "token": {},
                                                "expiredAt": {
                                                    "type": "integer"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/auth/company/logout": {
            "post": {
                "tags": [
                    "Company Authentication"
                ],
                "summary": "User logout",
                "description": "Invalidate access/refresh token pair for current session",
                "operationId": "90f2761b61b401c5bc442d80880ba23a",
                "responses": {
                    "200": {
                        "description": "Empty response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/auth/company/logout/all": {
            "post": {
                "tags": [
                    "Company Authentication"
                ],
                "summary": "Logout from all devices",
                "description": "Invalidate access/refresh token pairs for all user sessions",
                "operationId": "54da9ae718c0990c5e95af16c0a61900",
                "responses": {
                    "200": {
                        "description": "Empty response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {}
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
        "/api/auth/company/profile": {
            "get": {
                "tags": [
                    "Company Authentication"
                ],
                "summary": "Get user profile",
                "description": "Returns authenticated user's profile information",
                "operationId": "ec376210d25c351070e5bcf93f6186e0",
                "responses": {
                    "200": {
                        "description": "User profile information",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "name": {
                                            "type": "string",
                                            "example": "John Doe"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "user@example.com"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "JWT": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "name": "JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}